+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Use `SyncConfigType` qualifier for daily and frequent sync configs
 - Bump Kotlin to v1.5.21
 - Updated `ContactPatientBottomSheet` behaviour for supporting patients with and without phone number
+- Change `SyncInterval` from an enum to a data class
 - [In Progress: 13 Jul 2021] Load online lookup API if patient not found locally
 
 ### Features